requirements.txt
absl-py==2.1.0
aenum==3.1.15
alembic==1.18.4
annotated-types==0.7.0
ansi2html==1.9.2
attrs==23.2.0
bcrypt==5.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cmaes==0.12.0
colorlog==6.10.1
contourpy==1.3.2
cryptography==46.0.5
cycler==0.12.1
dash==2.12.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==4.4.2
exceptiongroup==1.3.1
Flask==2.2.5
fonttools==4.61.1
greenlet==3.3.1
idna==3.11
iniconfig==2.3.0
invoke==1.7.3
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.0
jsonschema==4.19.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lxml==5.2.1
Mako==1.1.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.6.0
mdurl==0.1.2
mock==3.0.5
narwhals==2.16.0
nest-asyncio==1.6.0
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.8.1.78
optuna==3.3.0
packaging==26.0
pandas==2.0.1
paramiko==3.5.1
pathlib2==2.3.6
pillow==10.2.0
plotly==5.20.0
pluggy==1.6.0
protobuf==3.19.6
psutil==6.1.1
pyaml==26.2.1
pycparser==3.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.19.2
PyNaCl==1.6.2
pyparsing==3.3.2
pytest==8.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==5.3
referencing==0.37.0
requests==2.32.5
retrying==1.4.2
rich==13.9.4
rpds-py==0.30.0
safetensors==0.4.3
scikit-learn==1.7.2
scikit-optimize==0.9.0
scipy==1.10.1
six==1.16.0
SQLAlchemy==2.0.46
tabulate==0.9.0
tenacity==9.1.4
threadpoolctl==3.6.0
tomli==2.4.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==2.2.3
XlsxWriter==1.2.2
